requirements.txt
python-dotenv == 1.0.0
dill == 0.3.7
langchain==0.0.284
python-dotenv==1.0.0
streamlit==1.22.0
unstructured==0.9.2
tiktoken==0.4.0
faiss-cpu==1.7.4
libmagic==1.0
OpenAI == 0.28.0
python-magic==0.4.27
